TGP-2890: move nav file
